refactor(cache-proposals): move entire feature to proprietary as Pro tier
Per review feedback, cache intelligence becomes a Pro feature gated by a
new CACHE_INTELLIGENCE entitlement.

Backend:
- Move apps/api/src/cache-proposals/ -> proprietary/cache-proposals/
- Module becomes @Global and is conditionally loaded via try/catch in
  app.module.ts, matching the inference-latency-pro pattern
- Extract MCP cache routes from mcp.controller.ts into a new proprietary
  CacheProposalMcpController so the routes simply don't exist when the
  proprietary module isn't loaded; community-tier deployments return 404
  on every cache endpoint and the MCP tools surface that to the agent
- HTTP and MCP controllers both gate on @UseGuards(LicenseGuard) +
  @RequiresFeature(Feature.CACHE_INTELLIGENCE), returning 402 when not
  entitled
- Extract shared MCP helpers (ValidateInstanceIdPipe, safeLimit, etc.)
  into apps/api/src/mcp/mcp-helpers.ts so both controllers share them
- Update internal imports in moved files to use @app/* aliases

Shared:
- Add Feature.CACHE_INTELLIGENCE under Tier.pro

Frontend:
- NavItem for /cache-proposals adds requiredFeature so the link locks
  for non-entitled users (matches Anomaly Detection pattern)
- usePendingProposals accepts an enabled flag; useCacheProposalsUnread
  short-circuits on non-entitled licenses so we don't poll every 15s and
  get 402s

Tests pass: api 1228/1235, web 175/175. No regressions vs. pre-move.
