+++ apps/api/src/app.module.ts
@@ -22,7 +22,6 @@ import { McpModule } from './mcp/mcp.module';
 import { MetricForecastingModule } from './metric-forecasting/metric-forecasting.module';
 import { InferenceLatencyModule } from './inference-latency/inference-latency.module';
 import { CliModule } from './cli/cli.module';
-import { CacheProposalsModule } from './cache-proposals/cache-proposals.module';
 
 let AiModule: any = null;
 let LicenseModule: any = null;
@@ -30,6 +29,7 @@ let KeyAnalyticsModule: any = null;
 let AnomalyModule: any = null;
 let WebhookProModule: any = null;
 let InferenceLatencyProModule: any = null;
+let CacheProposalsModule: any = null;
 let AgentModule: any = null;
 let DataRetentionModule: any = null;
 
@@ -78,6 +78,14 @@ try {
   const inferenceLatencyProModule = require('../../../proprietary/inference-latency-pro/inference-latency-pro.module');
   InferenceLatencyProModule = inferenceLatencyProModule.InferenceLatencyProModule;
   console.log('[InferenceLatencyPro] Proprietary module loaded');
+} catch {
+  // Proprietary module not available
+}
+
+try {
+  const cacheProposalsModule = require('../../../proprietary/cache-proposals/cache-proposals.module');
+  CacheProposalsModule = cacheProposalsModule.CacheProposalsModule;
+  console.log('[CacheProposals] Proprietary module loaded');
 } catch {
   // Proprietary module not available
 }
@@ -137,7 +145,6 @@ const baseImports = [
   MetricForecastingModule,
   InferenceLatencyModule,
   CliModule,
-  CacheProposalsModule,
 ];
 
 const proprietaryImports = [
@@ -146,6 +153,7 @@ const proprietaryImports = [
   AnomalyModule,
   WebhookProModule,
   InferenceLatencyProModule,
+  CacheProposalsModule,
   AiModule,
   AgentModule,
   DataRetentionModule,
