+++ apps/api/src/mcp/mcp.module.ts
@@ -7,7 +7,6 @@ import { CommandLogAnalyticsModule } from '../commandlog-analytics/commandlog-an
 import { ClientAnalyticsModule } from '../client-analytics/client-analytics.module';
 import { ClusterModule } from '../cluster/cluster.module';
 import { TelemetryModule } from '../telemetry/telemetry.module';
-import { CacheProposalsModule } from '../cache-proposals/cache-proposals.module';
 
 const logger = new Logger('McpModule');
 
@@ -42,7 +41,7 @@ const tokenProviders = AgentTokensServiceClass
 const optionalImports = [AnomalyModule].filter(Boolean);
 
 @Module({
-  imports: [StorageModule, MetricsModule, CommandLogAnalyticsModule, ClientAnalyticsModule, ClusterModule, TelemetryModule, CacheProposalsModule, ...optionalImports],
+  imports: [StorageModule, MetricsModule, CommandLogAnalyticsModule, ClientAnalyticsModule, ClusterModule, TelemetryModule, ...optionalImports],
   controllers: [McpController],
   providers: [AgentTokenGuard, ...tokenProviders],
 })
