+++ proprietary/cache-proposals/__tests__/cache-approval.service.spec.ts
@@ -1,4 +1,4 @@
-import { MemoryAdapter } from '../../storage/adapters/memory.adapter';
+import { MemoryAdapter } from '@app/storage/adapters/memory.adapter';
 import { CacheApplyService } from '../cache-apply.service';
 import { CacheProposalService } from '../cache-proposal.service';
 import { CacheResolverService, type ResolvedCache } from '../cache-resolver.service';
