+++ proprietary/cache-proposals/cache-readonly.types.ts
